Sep_25_Erica
I understand why you had the cursor hidden, but since I still need to do a lot of playtesting, I’ve made it visible again. I have a system where I need to click on objects!

The Cashier isn’t perfect yet, but the simulation of interacting with the NPC and entering input is mostly working! Just in case, please don’t touch my files or the things in the Hierarchy 😭